pkp/pkp-lib#7495 Hide new submission listing behind enable_new_submission_listing flag
